Sang
update bán hàng lần 2

+++ src/main/java/com/example/datn_trendsetter/Entity/SanPhamChiTiet.java
@@ -37,7 +37,7 @@ public class SanPhamChiTiet {
     private Integer soLuong;
 
     @Column(name = "gia")
-    private Double gia;
+    private Float gia;
 
     @OneToMany(mappedBy = "sanPhamChiTiet",fetch = FetchType.LAZY)
     private List<HinhAnh> hinhAnh;
